test(maestro): split the scorecard oracles into their own suite
